Do not write changes to /etc/fwupd/remotes.d
Write to a mutable version of the remote in /var/lib/fwupd/remotes.d
rather than writing to /etc/fwupd/remotes.d which may be immutable on
some systems. This also allows distributors to turn off features without
modifying the core fwupd package.

The values in /var/lib override the ones in /etc.